+++ src/main/java/ru/netology/Radio.java
@@ -4,17 +4,20 @@ public class Radio {
 
     private int currentRadioStation;
 
-
     private int maxStation;
 
+    private int currentVolume;
+    private int maxVolume;
 
 
     public Radio() {
         maxStation = 9;
+        maxVolume = 10;
     }
 
     public Radio(int stationsCount) {
         maxStation = stationsCount - 1;
+        maxVolume = stationsCount - 1;
 
     }
 
@@ -50,6 +53,47 @@ public class Radio {
         this.currentRadioStation = currentRadioStation;
     }
 
+
+
+
+    public void Volume(int volumeCount) {
+        maxVolume = volumeCount - 1;
+    }
+
+
+    public void nextCurrentVolume() {
+        if (currentVolume != maxVolume) {
+            currentVolume++;
+        } else {
+            currentVolume = maxVolume;
+        }
+
+    }
+
+    public void prevVolume() {
+        if (currentVolume != 0) {
+            currentVolume--;
+        } else {
+            currentVolume = 0;
+        }
+    }
+
+    public int getCurrentVolume() {
+        return currentVolume;
+    }
+
+    public void setCurrentVolume(int currentVolume) {
+        if (currentVolume < 0) {
+            return;
+        }
+        if (currentVolume > maxVolume) {
+            return;
+        }
+        this.currentVolume = currentVolume;
+    }
 }
+
+
+
